Fix IdList and make it not thread safe
